Adding new dragons data hook.
